Linux build fixes

+++ plugins/COMCORE/comcorepluginDBL/src/comcorepluginDBL_CDBLNetworkInterface.cpp
@@ -71,7 +71,7 @@ namespace DBL {
 //-------------------------------------------------------------------------*/
 
 bool
-CDBLNetworkInterface::SetupAdapterInfo( void* pAdaptInfoVoid ) 
+CDBLNetworkInterface::SetupAdapterInfo( void* pAdaptInfoVoid )
 {
 
 
@@ -80,9 +80,9 @@ CDBLNetworkInterface::SetupAdapterInfo( void* pAdaptInfoVoid )
 
 /*-------------------------------------------------------------------------*/
 
-bool 
-CDBLNetworkInterface::EnumNetworkAdapters( TINetworkInterfacePtrVector& interfaces ) 
-{	
+bool
+CDBLNetworkInterface::EnumNetworkAdapters( TINetworkInterfacePtrVector& interfaces )
+{
 
 	return true;
 }
@@ -93,11 +93,11 @@ CDBLNetworkInterface::CDBLNetworkInterface( void )
     : CINetworkInterface()
 	, m_dataLock()
 {GUCEF_TRACE;
-          
+
 }
 
 /*-------------------------------------------------------------------------*/
-    
+
 CDBLNetworkInterface::~CDBLNetworkInterface()
 {GUCEF_TRACE;
 
@@ -109,12 +109,12 @@ const CORE::CString&
 CDBLNetworkInterface::GetClassTypeName( void ) const
 {GUCEF_TRACE;
 
-    static const CORE::CString classTypeName = "GUCEF::COMCOREPLUGIN::DBL::::CDBLNetworkInterface"; 
+    static const CORE::CString classTypeName = "GUCEF::COMCOREPLUGIN::DBL::::CDBLNetworkInterface";
     return classTypeName;
 }
 
 /*-------------------------------------------------------------------------*/
-    
+
 CORE::CString
 CDBLNetworkInterface::GetCommunicationInterfaceType( void ) const
 {GUCEF_TRACE;
@@ -124,7 +124,7 @@ CDBLNetworkInterface::GetCommunicationInterfaceType( void ) const
 
 /*-------------------------------------------------------------------------*/
 
-CORE::CString 
+CORE::CString
 CDBLNetworkInterface::GetAdapterName( void ) const
 {GUCEF_TRACE;
 
@@ -134,7 +134,7 @@ CDBLNetworkInterface::GetAdapterName( void ) const
 
 /*-------------------------------------------------------------------------*/
 
-CORE::CString 
+CORE::CString
 CDBLNetworkInterface::GetAdapterDescription( void ) const
 {GUCEF_TRACE;
 
@@ -144,7 +144,7 @@ CDBLNetworkInterface::GetAdapterDescription( void ) const
 
 /*-------------------------------------------------------------------------*/
 
-UInt32 
+UInt32
 CDBLNetworkInterface::GetNrOfDnsAddresses( void ) const
 {GUCEF_TRACE;
 
@@ -153,8 +153,8 @@ CDBLNetworkInterface::GetNrOfDnsAddresses( void ) const
 }
 
 /*-------------------------------------------------------------------------*/
-    
-bool 
+
+bool
 CDBLNetworkInterface::GetDnsAddresses( THostAddressVector& dnsAddresses )
 {GUCEF_TRACE;
 
@@ -164,7 +164,7 @@ CDBLNetworkInterface::GetDnsAddresses( THostAddressVector& dnsAddresses )
 
 /*-------------------------------------------------------------------------*/
 
-UInt32 
+UInt32
 CDBLNetworkInterface::GetNrOfIPAddresses( void ) const
 {GUCEF_TRACE;
 
@@ -174,7 +174,7 @@ CDBLNetworkInterface::GetNrOfIPAddresses( void ) const
 
 /*-------------------------------------------------------------------------*/
 
-bool 
+bool
 CDBLNetworkInterface::GetIPInfo( TIPInfoVector& ipInfo, bool includeUninitialized ) const
 {GUCEF_TRACE;
 
@@ -183,8 +183,8 @@ CDBLNetworkInterface::GetIPInfo( TIPInfoVector& ipInfo, bool includeUninitialize
 }
 
 /*-------------------------------------------------------------------------*/
-		
-bool 
+
+bool
 CDBLNetworkInterface::IsDhcpUsed( void ) const
 {GUCEF_TRACE;
 
@@ -194,7 +194,7 @@ CDBLNetworkInterface::IsDhcpUsed( void ) const
 
 /*-------------------------------------------------------------------------*/
 
-CORE::CDateTime 
+CORE::CDateTime
 CDBLNetworkInterface::GetDhcpLeaseObtainedTime( void ) const
 {GUCEF_TRACE;
 
@@ -205,7 +205,7 @@ CDBLNetworkInterface::GetDhcpLeaseObtainedTime( void ) const
 
 /*-------------------------------------------------------------------------*/
 
-CORE::CDateTime 
+CORE::CDateTime
 CDBLNetworkInterface::GetDhcpLeaseExpirationTime( void ) const
 {GUCEF_TRACE;
 
@@ -216,7 +216,7 @@ CDBLNetworkInterface::GetDhcpLeaseExpirationTime( void ) const
 
 /*-------------------------------------------------------------------------*/
 
-bool 
+bool
 CDBLNetworkInterface::ReleaseAddress( void )
 {GUCEF_TRACE;
 
@@ -226,7 +226,7 @@ CDBLNetworkInterface::ReleaseAddress( void )
 
 /*-------------------------------------------------------------------------*/
 
-bool 
+bool
 CDBLNetworkInterface::RenewAddress( void )
 {GUCEF_TRACE;
 
@@ -236,7 +236,7 @@ CDBLNetworkInterface::RenewAddress( void )
 
 /*-------------------------------------------------------------------------*/
 
-bool 
+bool
 CDBLNetworkInterface::IsWinsUsed( void ) const
 {GUCEF_TRACE;
 
@@ -256,7 +256,7 @@ CDBLNetworkInterface::GetPrimaryWinsServer( void ) const
 
 /*-------------------------------------------------------------------------*/
 
-COMCORE::CIPv4Address 
+COMCORE::CIPv4Address
 CDBLNetworkInterface::GetSecondaryWinsServer( void ) const
 {GUCEF_TRACE;
 
@@ -266,7 +266,7 @@ CDBLNetworkInterface::GetSecondaryWinsServer( void ) const
 
 /*-------------------------------------------------------------------------*/
 
-UInt32 
+UInt32
 CDBLNetworkInterface::GetOsAdapterIndex( void ) const
 {GUCEF_TRACE;
 
@@ -276,7 +276,7 @@ CDBLNetworkInterface::GetOsAdapterIndex( void ) const
 
 /*-------------------------------------------------------------------------*/
 
-bool 
+bool
 CDBLNetworkInterface::GetMetrics( COMCORE::CNetworkInterfaceMetrics& metrics ) const
 {GUCEF_TRACE;
 
@@ -290,7 +290,7 @@ CDBLNetworkInterface::GetMetrics( COMCORE::CNetworkInterfaceMetrics& metrics ) c
 	int32_t productLookupResult = dblwrapper_mal_translate_to_product( nicIndex, &productIdPartA, &productIdPartB );
 	if ( 0 != productLookupResult )
 	{
-		GUCEF_ERROR_LOG( CORE::LOGLEVEL_IMPORTANT, "DBLNetworkInterface: Couldn't find board " + CORE::ToString( nicIndex ) );
+		GUCEF_ERROR_LOG( CORE::LOGLEVEL_IMPORTANT, "DBLNetworkInterface: Couldn't find board " + CORE::ToString( (Int64)nicIndex ) );
 		return false;
 	}
 
@@ -298,26 +298,26 @@ CDBLNetworkInterface::GetMetrics( COMCORE::CNetworkInterfaceMetrics& metrics ) c
 	int64_t v8 = 0;
 	int32_t optionFlags = v8 & 0xFFFFFFFFFFFFFF00 | productIdPartA;
 	void* deviceContext = GUCEF_NULL;
-	int32_t openResult = dblwrapper_mal_open( productIdPartB, 
-										      0xffffffff, 
-											  &deviceContext, 
-											  optionFlags, 
+	int32_t openResult = dblwrapper_mal_open( productIdPartB,
+										      0xffffffff,
+											  &deviceContext,
+											  optionFlags,
 											  productIdPartB );
 	if ( 0 != openResult )
 	{
-		GUCEF_ERROR_LOG( CORE::LOGLEVEL_IMPORTANT, "DBLNetworkInterface: Cannot open board " + CORE::ToString( nicIndex ) );
+		GUCEF_ERROR_LOG( CORE::LOGLEVEL_IMPORTANT, "DBLNetworkInterface: Cannot open board " + CORE::ToString( (Int64)nicIndex ) );
 		return false;
 	}
 
 	// Query the numeric NIC ID first as a sanity check to check our access
 	int32_t driverId = productIdPartB; // <- on input we need to give the product id for the board we are asking for
-	int32_t queryResult = dblwrapper_mal_ioctl( deviceContext, 
-												MAL_IOCTL_INFOTYPEID_NIC, 
-												&driverId, 
+	int32_t queryResult = dblwrapper_mal_ioctl( deviceContext,
+												MAL_IOCTL_INFOTYPEID_NIC,
+												&driverId,
 												MAL_IOCTL_PROPERTYID_INT32_ID );
 	if ( 0 != queryResult )
 	{
-		GUCEF_ERROR_LOG( CORE::LOGLEVEL_IMPORTANT, "DBLNetworkInterface: Failed to get driver id for board " + CORE::ToString( nicIndex ) );
+		GUCEF_ERROR_LOG( CORE::LOGLEVEL_IMPORTANT, "DBLNetworkInterface: Failed to get driver id for board " + CORE::ToString( (Int64)nicIndex ) );
 		dblwrapper_mal_close( deviceContext );
 		return false;
 	}
@@ -329,14 +329,14 @@ CDBLNetworkInterface::GetMetrics( COMCORE::CNetworkInterfaceMetrics& metrics ) c
 
 	// Query how many ports the NIC has, it can have multiple ports on a single card
 	int32_t	boardPortCount = productIdPartB; // <- on input we need to give the product id for the board we are asking for
-	queryResult = dblwrapper_mal_ioctl( deviceContext, 
-									    MAL_IOCTL_INFOTYPEID_PORTS, 
-										&boardPortCount, 
+	queryResult = dblwrapper_mal_ioctl( deviceContext,
+									    MAL_IOCTL_INFOTYPEID_PORTS,
+										&boardPortCount,
 										MAL_IOCTL_PROPERTYID_INT32_COUNT );
 	if ( 0 != queryResult )
 	{
 		// we wont count this as a fatal error. log but keep going
-		GUCEF_WARNING_LOG( CORE::LOGLEVEL_IMPORTANT, "DBLNetworkInterface: Cannot get port count for board " + CORE::ToString( nicIndex ) );
+		GUCEF_WARNING_LOG( CORE::LOGLEVEL_IMPORTANT, "DBLNetworkInterface: Cannot get port count for board " + CORE::ToString( (Int64)nicIndex ) );
 	}
 
 	GUCEF_SYSTEM_LOG( CORE::LOGLEVEL_NORMAL, "DBLNetworkInterface: Board NIC ID string is \"" + CORE::ToString( idStr ) +
@@ -344,13 +344,13 @@ CDBLNetworkInterface::GetMetrics( COMCORE::CNetworkInterfaceMetrics& metrics ) c
 
 	// Query how many counter stats are available, we will need to allocate sufficient storage for them
 	int32_t nrOfCounterStats = productIdPartB; // <- on input we need to give the product id for the board we are asking for
-	queryResult = dblwrapper_mal_ioctl( deviceContext, 
-										MAL_IOCTL_INFOTYPEID_COUNTERS_METADATA, 
-										&nrOfCounterStats, 
+	queryResult = dblwrapper_mal_ioctl( deviceContext,
+										MAL_IOCTL_INFOTYPEID_COUNTERS_METADATA,
+										&nrOfCounterStats,
 										MAL_IOCTL_PROPERTYID_INT32_COUNT );
 	if ( 0 != queryResult )
 	{
-		GUCEF_ERROR_LOG( CORE::LOGLEVEL_IMPORTANT, "DBLNetworkInterface: Failed to get nr of counter stats for board " + CORE::ToString( nicIndex ) );
+		GUCEF_ERROR_LOG( CORE::LOGLEVEL_IMPORTANT, "DBLNetworkInterface: Failed to get nr of counter stats for board " + CORE::ToString( (Int64)nicIndex ) );
 		dblwrapper_mal_close( deviceContext );
 		return false;
 	}
@@ -370,20 +370,20 @@ CDBLNetworkInterface::GetMetrics( COMCORE::CNetworkInterfaceMetrics& metrics ) c
 		 GUCEF_NULL == counters2   ||
 		 GUCEF_NULL == counters3    )
 	{
-		GUCEF_ERROR_LOG( CORE::LOGLEVEL_IMPORTANT, "DBLNetworkInterface: Failed to allocate memory to store counter stats for board " + CORE::ToString( nicIndex ) );
+		GUCEF_ERROR_LOG( CORE::LOGLEVEL_IMPORTANT, "DBLNetworkInterface: Failed to allocate memory to store counter stats for board " + CORE::ToString( (Int64)nicIndex ) );
 		dblwrapper_mal_close( deviceContext );
 		return false;
 	}
 
  	// Query the counter strings
 	counterStrsBuffer.AsType< int32_t >() = productIdPartB;	// <- on input we need to give the product id for the board we are asking for
-	queryResult = dblwrapper_mal_ioctl( deviceContext, 
-										MAL_IOCTL_INFOTYPEID_COUNTERS_STRINGS, 
-										counterStrs, 
+	queryResult = dblwrapper_mal_ioctl( deviceContext,
+										MAL_IOCTL_INFOTYPEID_COUNTERS_STRINGS,
+										counterStrs,
 										MAL_IOCTL_PROPERTYID_INT32_COUNT );	// <- ???
 	if ( 0 != queryResult )
 	{
-		GUCEF_ERROR_LOG( CORE::LOGLEVEL_IMPORTANT, "DBLNetworkInterface: Failed to get nr of counter stats for board " + CORE::ToString( nicIndex ) );
+		GUCEF_ERROR_LOG( CORE::LOGLEVEL_IMPORTANT, "DBLNetworkInterface: Failed to get nr of counter stats for board " + CORE::ToString( (Int64)nicIndex ) );
 		dblwrapper_mal_close( deviceContext );
 		return false;
 	}
@@ -394,14 +394,14 @@ CDBLNetworkInterface::GetMetrics( COMCORE::CNetworkInterfaceMetrics& metrics ) c
 	counters2Buffer.AsType< int32_t >() = productIdPartB;	// <- on input we need to give the product id for the board we are asking for
     int32_t counterQueryType = 0;
 	queryResult = dblwrapper_mal_ioctl( deviceContext, 2236563, counters2, MAL_IOCTL_PROPERTYID_INT32_COUNT );
-	if( (uint32_t)queryResult == 0 ) 
+	if( (uint32_t)queryResult == 0 )
 	{
 		*v14 = productIdPartB;
 		v18 = 1;
 		v19 = v14;
 		counterQueryType = MAL_IOCTL_INFOTYPEID_COUNTERS;
-	} 
-	else 
+	}
+	else
 	{
 		GUCEF_DEBUG_LOG( CORE::LOGLEVEL_NORMAL, "DBLNetworkInterface: no classes support" );
 		*v13 = productIdPartB;
@@ -416,44 +416,44 @@ CDBLNetworkInterface::GetMetrics( COMCORE::CNetworkInterfaceMetrics& metrics ) c
 
 
         queryResult = dblwrapper_mal_ioctl( deviceContext, counterQueryType, v19, MAL_IOCTL_PROPERTYID_INT32_COUNT );
-        if( (uint32_t)queryResult != 0 ) 
+        if( (uint32_t)queryResult != 0 )
 		{
             _ErrMsg = "Cannot get counters values";
-        } 
-		else 
+        }
+		else
 		{
-            if( local_0x90 != 0 ) 
+            if( local_0x90 != 0 )
 			{
                 v22 = local_0x90;
                 v23 = v13;
-                while( 1 ) 
+                while( 1 )
 				{
-                    if( local_0x74 < 2 ) 
+                    if( local_0x74 < 2 )
 					{
                         v25 = strstr( _Str, "(Port 1)" );
-                        if( v25 == 0 ) 
+                        if( v25 == 0 )
 						{
                             goto node_244;
                         }
-                    } 
-					else 
+                    }
+					else
 					{
                         node_244:
-                        if( (int32_t)v18 == 0 ) 
+                        if( (int32_t)v18 == 0 )
 						{
-                            if( ((int32_t)v3 == 0 || (int32_t)v18 != *v23) && ((int32_t)v2 != 0 || (uint32_t)*(v15 - v13 + v23) == 0) ) 
+                            if( ((int32_t)v3 == 0 || (int32_t)v18 != *v23) && ((int32_t)v2 != 0 || (uint32_t)*(v15 - v13 + v23) == 0) )
 							{
                                 v24 = ntohl( *v23 );
                                 printf( "%34s: %10u\n", _Str, v24 );
                             }
-                        } 
-						else 
-						if( ((int32_t)v3 == 0 || *v14 != 0) && ((int32_t)v2 != 0 || (uint32_t)*(v15 - v13 + v23) == 0) ) 
+                        }
+						else
+						if( ((int32_t)v3 == 0 || *v14 != 0) && ((int32_t)v2 != 0 || (uint32_t)*(v15 - v13 + v23) == 0) )
 						{
                             printf( "%34s: %20llu\n", _Str, *v14 );
                         }
                     }
-                    if( v22 == 1 ) 
+                    if( v22 == 1 )
 					{
                         break;
                     }
@@ -463,27 +463,27 @@ CDBLNetworkInterface::GetMetrics( COMCORE::CNetworkInterfaceMetrics& metrics ) c
                     (uint8_t *)v23 += 4;
                 }
             }
-            if( local_0x78 != 0 ) 
+            if( local_0x78 != 0 )
 			{
                 local_0x70 = local_0x98;
                 queryResult = dblwrapper_mal_ioctl( deviceContext, MAL_IOCTL_INFOTYPEID_IRQ, &local_0x70, 16 );
-                if( (uint32_t)queryResult == 0 ) 
+                if( (uint32_t)queryResult == 0 )
 				{
                     printf( "Host irq info: %10u intr, %10u events, %10u intr w/o event\n", local_0x6C, local_0x68, local_0x64 );
-                } 
-				else 
+                }
+				else
 				{
                     perror( "Cannot get interrupt counters -- try enabling MAL_DEBUG in the kernel?" );
                 }
             }
-            if( local_0x80 == 0 ) 
+            if( local_0x80 == 0 )
 			{
                 dblwrapper_mal_close( deviceContext );
                 __security_check_cookie( local_0x30 ^ &v1 );
                 return 0;
             }
             queryResult = dblwrapper_mal_ioctl( deviceContext, MAL_IOCTL_INFOTYPEID_CLEAR_COUNTERS, &local_0x98, 4 );
-            if( (uint32_t)queryResult == 0 ) 
+            if( (uint32_t)queryResult == 0 )
 			{
                 dblwrapper_mal_close( local_0x88 );
                 __security_check_cookie( local_0x30 ^ &v1 );
@@ -499,7 +499,7 @@ CDBLNetworkInterface::GetMetrics( COMCORE::CNetworkInterfaceMetrics& metrics ) c
 
 /*-------------------------------------------------------------------------*/
 
-const MT::CILockable* 
+const MT::CILockable*
 CDBLNetworkInterface::AsLockable( void ) const
 {GUCEF_TRACE;
